fixed template to build layer
