+++ deploy/bottle_api.py
@@ -11,17 +11,16 @@ import tempfile
 import base64
 import functools
 import io
-#import urllib
 import csv
 import os
 import zipfile
 from collections import defaultdict
 from zipfile import ZipFile
 
-from validate_email_address import validate_email
 import bottle
 from bottle import request,response
-from zappa.asynchronous import task
+from validate_email_address import validate_email
+#from zappa.asynchronous import task
 
 import db
 import db_object
@@ -154,6 +153,90 @@ def get_user_dict():
         auth.clear_cookie()
         raise auth.http403(f'Error 404: api_key {api_key} is invalid. ')
     return userdict
+
+def page_dict(title='', *, require_auth=False, lookup=True, logout=False,debug=False):
+    """Returns a dictionary that can be used by post of the templates.
+    :param: title - the title we should give the page
+    :param: require_auth  - if true, the user must already be authenticated, or throws an error
+    :param: logout - if true, force the user to log out by issuing a clear-cookie command
+    :param: lookup - if true, we weren't being called in an error condition, so we can lookup the api_key in the URL or the cookie
+    """
+    logging.debug("1. page_dict require_auth=%s logout=%s lookup=%s",require_auth,logout,lookup)
+    o = urlparse(request.url)
+    logging.debug("o=%s",o)
+    if lookup:
+        api_key = auth.get_user_api_key()
+        logging.debug("auth.get_user_api_key=%s",api_key)
+        if api_key is None and require_auth is True:
+            logging.debug("api_key is None and require_auth is True")
+            raise auth.http403("api_key is None and require_auth is True")
+    else:
+        api_key = None
+
+    if (api_key is not None) and (logout is False):
+        # Get the user_dict is from the database
+        user_dict = get_user_dict()
+        user_name = user_dict['name']
+        user_email = user_dict['email']
+        user_demo  = user_dict['demo']
+        user_id = user_dict['id']
+        user_primary_course_id = user_dict['primary_course_id']
+        logged_in = 1
+        primary_course_name = db.lookup_course_by_id(course_id=user_primary_course_id)['course_name']
+        admin = 1 if db.check_course_admin(user_id=user_id, course_id=user_primary_course_id) else 0
+        # If this is a demo account, the user cannot be an admin (security)
+        if user_demo:
+            assert not admin
+        # if this is not a demo account and the user_id is set, make sure we set the cookie
+        if (user_id is not None) and (not user_demo):
+            auth.set_cookie(api_key)
+
+    else:
+        user_name  = None
+        user_email = None
+        user_demo  = 0
+        user_id    = None
+        user_primary_course_id = None
+        primary_course_name = None
+        admin = 0
+        logged_in = 0
+
+    try:
+        movie_id = int(request.query.get('movie_id'))
+    except (AttributeError, KeyError, TypeError):
+        movie_id = 0            # to avoid errors
+
+    ret= fix_types({
+        C.API_BASE: api_base,
+        C.STATIC_BASE: static_base,
+        'api_key': api_key,     # the API key that is currently active
+        'user_id': user_id,     # the user_id that is active
+        'user_name': user_name, # the user's name
+        'user_email': user_email, # the user's email
+        'user_demo':  user_demo,  # true if this is a demo user
+        'logged_in': logged_in,
+        'admin':admin,
+        'user_primary_course_id': user_primary_course_id,
+        'primary_course_name': primary_course_name,
+        'title':'Plant Tracer '+title,
+        'hostname':o.hostname,
+        'movie_id':movie_id,
+        'demo_mode_available':DEMO_MODE_AVAILABLE,
+        'MAX_FILE_UPLOAD': C.MAX_FILE_UPLOAD,
+        'dbreader_host':auth.get_dbreader().host,
+        'version':__version__,
+        'git_head_time':git_head_time(),
+        'git_last_commit':git_last_commit()
+    })
+    for (k,v) in ret.items():
+        if v is None:
+            ret[k] = "null"
+    if debug:
+        logging.debug("fixed page_dict=%s",ret)
+    return ret
+
+
+
 
 ################################################################
 # /api URLs
@@ -727,8 +810,10 @@ class MovieTrackCallback:
 
 ##
 ## @task causes this to be run in background on zappa, but in foreground when run locally
+## It's specific to zappa
+## TODO - replace with a reasonable background task, or give up on background tasks
 ##
-@task
+#@task
 def api_track_movie(*,user_id, movie_id, frame_start):
     """Generate trackpoints for a movie based on initial trackpoints stored in the database at frame_start.
     Stores new trackpoints and each frame in the database. No longer renders new movie: that's now in render_tracked_movie
