+++ deploy/bottle_app.py
@@ -53,9 +53,9 @@ from bottle import request
 
 # Bottle creates a large number of no-member errors, so we just remove the warning
 # pylint: disable=no-member
-from lib.ctools import clogging
+import clogging
 
-import wsgiserver               # pylint: disable=syntax-error
+#import wsgiserver               # pylint: disable=syntax-error
 import db
 import db_object
 import auth
@@ -64,8 +64,6 @@ from auth import DEMO_MODE_AVAILABLE
 from paths import view, STATIC_DIR
 from constants import C,__version__,GET,GET_POST
 
-import bottle_api
-from bottle_api import git_head_time,git_last_commit,get_user_dict,fix_types
 import dbmaint
 
 DEFAULT_OFFSET = 0
@@ -87,8 +85,17 @@ static_base = os.getenv(C.PLANTTRACER_STATIC_BASE,'/')
 ### STARTUP CODE RUNS WHEN THIS IS LOADED ###
 #############################################
 
-app = bottle.default_app()      # for Lambda
-app.mount('/api', bottle_api.api)
+################
+## API SUPPORT
+## Commented out for now
+
+# import bottle_api
+# from bottle_api import page_dict
+# app = bottle.default_app()      # for Lambda
+# app.mount('/api', bottle_api.api)
+def page_dict():
+    return {}
+
 
 def fix_boto_log_level():
     for name in logging.root.manager.loggerDict:
@@ -150,88 +157,6 @@ def favicon():
 ################################################################
 # HTML Pages served with template system
 ################################################################
-
-def page_dict(title='', *, require_auth=False, lookup=True, logout=False,debug=False):
-    """Returns a dictionary that can be used by post of the templates.
-    :param: title - the title we should give the page
-    :param: require_auth  - if true, the user must already be authenticated, or throws an error
-    :param: logout - if true, force the user to log out by issuing a clear-cookie command
-    :param: lookup - if true, we weren't being called in an error condition, so we can lookup the api_key in the URL or the cookie
-    """
-    logging.debug("1. page_dict require_auth=%s logout=%s lookup=%s",require_auth,logout,lookup)
-    o = urlparse(request.url)
-    logging.debug("o=%s",o)
-    if lookup:
-        api_key = auth.get_user_api_key()
-        logging.debug("auth.get_user_api_key=%s",api_key)
-        if api_key is None and require_auth is True:
-            logging.debug("api_key is None and require_auth is True")
-            raise auth.http403("api_key is None and require_auth is True")
-    else:
-        api_key = None
-
-    if (api_key is not None) and (logout is False):
-        # Get the user_dict is from the database
-        user_dict = get_user_dict()
-        user_name = user_dict['name']
-        user_email = user_dict['email']
-        user_demo  = user_dict['demo']
-        user_id = user_dict['id']
-        user_primary_course_id = user_dict['primary_course_id']
-        logged_in = 1
-        primary_course_name = db.lookup_course_by_id(course_id=user_primary_course_id)['course_name']
-        admin = 1 if db.check_course_admin(user_id=user_id, course_id=user_primary_course_id) else 0
-        # If this is a demo account, the user cannot be an admin (security)
-        if user_demo:
-            assert not admin
-        # if this is not a demo account and the user_id is set, make sure we set the cookie
-        if (user_id is not None) and (not user_demo):
-            auth.set_cookie(api_key)
-
-    else:
-        user_name  = None
-        user_email = None
-        user_demo  = 0
-        user_id    = None
-        user_primary_course_id = None
-        primary_course_name = None
-        admin = 0
-        logged_in = 0
-
-    try:
-        movie_id = int(request.query.get('movie_id'))
-    except (AttributeError, KeyError, TypeError):
-        movie_id = 0            # to avoid errors
-
-    ret= fix_types({
-        C.API_BASE: api_base,
-        C.STATIC_BASE: static_base,
-        'api_key': api_key,     # the API key that is currently active
-        'user_id': user_id,     # the user_id that is active
-        'user_name': user_name, # the user's name
-        'user_email': user_email, # the user's email
-        'user_demo':  user_demo,  # true if this is a demo user
-        'logged_in': logged_in,
-        'admin':admin,
-        'user_primary_course_id': user_primary_course_id,
-        'primary_course_name': primary_course_name,
-        'title':'Plant Tracer '+title,
-        'hostname':o.hostname,
-        'movie_id':movie_id,
-        'demo_mode_available':DEMO_MODE_AVAILABLE,
-        'MAX_FILE_UPLOAD': C.MAX_FILE_UPLOAD,
-        'dbreader_host':auth.get_dbreader().host,
-        'version':__version__,
-        'git_head_time':git_head_time(),
-        'git_last_commit':git_last_commit()
-    })
-    for (k,v) in ret.items():
-        if v is None:
-            ret[k] = "null"
-    if debug:
-        logging.debug("fixed page_dict=%s",ret)
-    return ret
-
 
 @bottle.route('/', method=GET_POST)
 @view('index.html')
